clear_run now doesn't fail if run does not exist

+++ src/data/interim/clear_run.py
@@ -8,9 +8,9 @@ from . import get_path
 curpath = get_path()
 
 
-
 def clear_run(run):
     rmpath = os.path.join(curpath, run)
-    log.info("removing %s", rmpath)
-    shutil.rmtree(rmpath)
+    if os.path.isdir(rmpath):
+        log.info("removing %s", rmpath)
+        shutil.rmtree(rmpath)
